+++ src/main/java/de/dfki/grave/editor/EditorComponent.java
@@ -8,7 +8,7 @@ import java.util.Observer;
 import javax.swing.JComponent;
 
 import de.dfki.grave.editor.panels.WorkSpace;
-import de.dfki.grave.model.flow.geom.Position;
+import de.dfki.grave.model.flow.Position;
 import de.dfki.grave.model.project.EditorConfig;
 
 @SuppressWarnings("serial")
