+++ src/main/java/de/dfki/grave/editor/action/ModifyEdgeAction.java
@@ -5,7 +5,7 @@ import java.awt.Point;
 import de.dfki.grave.editor.panels.WorkSpace;
 import de.dfki.grave.model.flow.AbstractEdge;
 import de.dfki.grave.model.flow.BasicNode;
-import de.dfki.grave.model.flow.geom.Position;
+import de.dfki.grave.model.flow.Position;
 
 /**
  * @author Bernd Kiefer
