+++ src/main/java/de/dfki/grave/editor/action/MoveNodesAction.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import de.dfki.grave.editor.panels.WorkSpace;
 import de.dfki.grave.model.flow.BasicNode;
-import de.dfki.grave.model.flow.geom.Position;
+import de.dfki.grave.model.flow.Position;
 
 /**
  * @author Gregor Mehlmann
