+++ src/main/java/de/dfki/grave/editor/panels/WorkSpace.java
@@ -22,9 +22,6 @@ import de.dfki.grave.editor.event.ElementSelectedEvent;
 import de.dfki.grave.editor.event.ProjectChangedEvent;
 import de.dfki.grave.editor.event.WorkSpaceSelectedEvent;
 import de.dfki.grave.model.flow.*;
-import de.dfki.grave.model.flow.geom.Boundary;
-import de.dfki.grave.model.flow.geom.Geom;
-import de.dfki.grave.model.flow.geom.Position;
 import de.dfki.grave.model.project.EditorConfig;
 import de.dfki.grave.model.project.EditorProject;
 import de.dfki.grave.util.Pair;
