+++ src/main/java/de/dfki/grave/model/flow/Geom.java
@@ -1,4 +1,4 @@
-package de.dfki.grave.model.flow.geom;
+package de.dfki.grave.model.flow;
 
 import java.awt.Point;
 import java.awt.geom.Point2D;
@@ -54,31 +54,6 @@ public class Geom {
     Point result = new Point(a);
     result.translate(b.x, b.y);
     return result;
-  }
-
-  /*
-  public static Point add(Point a, Position b) {
-    Point result = new Point(a);
-    result.translate(b.getXPos(), b.getYPos());
-    return result;
-  }
-
-  // returns an angle in the range 0 .. 2*pi, where (0, 1) returns zero, and
-  // (1,0) returns pi/2
-  public static double angle(Point center, Point p) {
-    double res = Math.atan2(p.x - center.x, p.y - center.y);
-    if (res < 0) res = 2 * Math.PI + res;
-    return res;
-  }
-  */
-  
-  // returns an angle in the range 0 .. 2*pi, where (0, 1) returns zero, and
-  // (1,0) returns pi/2
-  public static double angle(Position center, Position p) { 
-    double res = Math.atan2(p.getXPos() - center.getXPos(), 
-        p.getYPos() - center.getYPos());
-    if (res < 0) res = 2 * Math.PI + res;
-    return res;
   }
 
   public static double dotProd(Point a, Point b) {
