+++ src/test/java/de/dfki/grave/GeomTest.java
@@ -7,7 +7,7 @@ import java.awt.geom.Point2D;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import de.dfki.grave.model.flow.geom.Geom;
+import de.dfki.grave.model.flow.Geom;
 
 public class GeomTest {
   @BeforeClass
